Revert "样式修改"
This reverts commit 6e896c681ec40e4eceee8292530728b75cd3f1a1.

+++ dolphinscheduler-ui/src/common/common.ts
@@ -46,7 +46,8 @@ export const bytesToSize = (bytes: number) => {
   const k = 1024 // or 1024
   const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
   const i = Math.floor(Math.log(bytes) / Math.log(k))
-  return parseFloat((bytes / Math.pow(k, i)).toPrecision(3)) + ' ' + sizes[i]
+
+  return parseInt((bytes / Math.pow(k, i)).toPrecision(3)) + ' ' + sizes[i]
 }
 
 export const fileTypeArr = [
